content/requirements.txt
openai
python-dotenv
anthropic
requests
pydantic
argparse
